+++ src/import_dats.rs
@@ -575,6 +575,8 @@ pub async fn reimport_orphan_romfiles(
             &header,
             &Path::new(&romfile.path),
             hash_algorithm,
+            true,
+            false,
         )
         .await?;
     }