+++ src/import_roms/test_chd_multiple_tracks_without_cue_should_fail.rs
@@ -47,6 +47,7 @@ async fn test() {
         &None,
         &romfile_path,
         &HashAlgorithm::Crc,
+        true,
     )
     .await
     .unwrap();