+++ src/import_roms/test_chd_single_track.rs
@@ -49,6 +49,7 @@ async fn test() {
         &None,
         &romfile_path,
         &HashAlgorithm::Crc,
+        true,
     )
     .await
     .unwrap();